TESTS: Refactor Maxwell unit tests
