+++ tests/unit/test_maxwell_3d.py
@@ -30,8 +30,16 @@ from ansys.aedt.core.maxwell import Maxwell3d
 import pytest
 
 
+@pytest.fixture
+def maxwell_3d_setup():
+    """Fixture used to mock the creation of a Maxwell instance."""
+    with patch("ansys.aedt.core.maxwell.Maxwell3d.__init__", lambda x: None):
+        mock_instance = MagicMock(spec=Maxwell3d)
+        yield mock_instance
+
+
 @patch.object(Maxwell3d, "solution_type", "Transient")
-@patch("ansys.aedt.core.maxwell.BoundaryObject", autospec=True)
+@patch("ansys.aedt.core.mixins.BoundaryObject", autospec=True)
 def test_maxwell_3d_assign_resistive_sheet_failure(mock_boundary_object, maxwell_3d_setup):
     boundary_object = MagicMock()
     boundary_object.create.return_value = False
@@ -40,5 +48,5 @@ def test_maxwell_3d_assign_resistive_sheet_failure(mock_boundary_object, maxwell
     maxwell._modeler = MagicMock()
     maxwell._logger = MagicMock()
 
-    with pytest.raises(AEDTRuntimeError, match=r"Boundary ResistiveSheet_\w+ was not created\."):
+    with pytest.raises(AEDTRuntimeError, match="Failed to create boundary"):
         maxwell.assign_resistive_sheet(None, None)
